Fixes #3907, move Vert.x producers to Resteasy
Resteasy will always run on top of vert.x in the new model
so this dependency arrangement makes more sense
