+++ extensions/resteasy-server-common/runtime/src/main/resources/META-INF/services/javax.ws.rs.ext.Providers
@@ -1,0 +1,4 @@
+io.quarkus.resteasy.server.common.vertx.JsonObjectWriter
+io.quarkus.resteasy.server.common.vertx.JsonArrayWriter
+io.quarkus.resteasy.server.common.vertx.JsonObjectReader
+io.quarkus.resteasy.server.common.vertx.JsonArrayReader
