--- extensions/vertx/runtime/src/main/resources/META-INF/services/javax.ws.rs.ext.Providers
@@ -1,4 +1,0 @@
-io.quarkus.vertx.runtime.JsonObjectWriter
-io.quarkus.vertx.runtime.JsonArrayWriter
-io.quarkus.vertx.runtime.JsonObjectReader
-io.quarkus.vertx.runtime.JsonArrayReader
